requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
pyopenssl
whois
python_whois
tldextract
bs4
pyppeteer